refactor methods
refactor methods
